fix: move project to the right folder and some fixes from code review
