+++ src/HealthChecks.Gcp.CloudFirestore/CloudFirestoreOptions.cs
